Update renaming_aps.py
The current codebase sends a 200 code when names with spaces are used but does not handle them properly and the new AP name does not save properly and is not shown in Central.

This allows for spaces to be used in AP names. I have validated it with Central AOS10.